Add rpm-file-permissions
As we are COPYing in some directories that are provided by the filesystem
rpm package we are unfortunately changing default permissions set on those
directories. This shell script uses information provided by an rpm query
to reset permissions of directories we are changing to their default values.

Fixes: rhbz#1481808

Also moves some commands in Dockerfiles to achieve better cache usage.