build_and_push: detect fedora tag from Dockerfile
Detect automatically fedora tag from Dockerfile in build_and_push_action.
The goal is to automatically detect change between Fedora version, e.g.,
35->36.

The only place, from where version transition can be spotted is
Dockerfile.fedora.

The used tag has to be in format "36".

curl for getting the Dockerfile.fedora was used because we do not have
repository cloned at that point in the action.
Cloning whole repository because of one Dockerfile would be ineffective.