bump build-and-push action version, remove deprecated set-output (#265)
New version of build-and-push changes approach to suffix of the
image. The suffix is no longer an input, instead image name input was
introduced. As this action run build-and-push for either s2i-core and s2i-base
image, the image name can not be set directly in the matrix. For that
reason the image name is constructed when calling build-and-push with
prefix and suffix. Hence the added dash in the suffix string.
The suffix is not relevant for Fedora images.