enable builds of core and base images for 2 Fedora releases
This commit brings support for f38 and f39, later f38 will be migrated to f40.
This change enables us to test on bleeding edge Fedora and sustain the
stable one for standard container images at the same time.

For backwards compatibilty also the Fedora target stays in form of
symlink to stable Fedora Dockerfile - currently f38